Absorb SplashScreenComponent into SplashScreen
