+++ src/Explorer/SplashScreen/SplashScreen.less
